tradingview_alpaca/requirements.txt
fastapi>=0.111.0
uvicorn[standard]
alpaca-py>=0.26.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv
httpx
